+++ src/prefabs/radioinput.tsx
@@ -316,7 +316,7 @@ const beforeCreate = ({
                 {
                   id: propertyId,
                   type: 'PROPERTY',
-                  name: `{{ ${model?.name}.${name} }}`,
+                  useKey: 'uuid',
                 },
               ];